requirements.txt
flask==2.3.3
mxnet==1.9.1
Pillow==10.0.0
numpy==1.24.3

